requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.8
feedparser==6.0.11
frozenlist==1.5.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
